feat: Rope Manager 구현
Mission을 상속 받아 해님달님 미니게임의 성공 여부 및 진행 상황을 초기화 해주는 역할을 한다
