+++ Assets/PrivateFolder/HyunJune/Prefab/RopeMission.prefab
@@ -3598,6 +3598,7 @@ GameObject:
   - component: {fileID: 3714933895854488849}
   - component: {fileID: 3714933895854488855}
   - component: {fileID: 3714933895854488848}
+  - component: {fileID: 2873271678099697791}
   m_Layer: 5
   m_Name: RopeMission
   m_TagString: Untagged
@@ -3668,6 +3669,19 @@ MonoBehaviour:
   m_FillOrigin: 0
   m_UseSpriteMesh: 0
   m_PixelsPerUnitMultiplier: 1
+--- !u!114 &2873271678099697791
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 3714933895854488850}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: e0add1bb95944f249858fbc71f5afc77, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  control: {fileID: 7603277285812282783}
 --- !u!1 &3714933897178467329
 GameObject:
   m_ObjectHideFlags: 0
