+++ Assets/PrivateFolder/HyunJune/Script/RopeGame/RopeManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e0add1bb95944f249858fbc71f5afc77
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
